Cambie la cola de receiver a Action*, arregle errores de memory leak en game, y errores de invalid read en receive_thread y send_thread. Tambien elimine punteros innecesarios en client_accepter y server_client. Hay errores en la cola de receiver (invalid reads).

+++ server/action_moving.cpp
@@ -2,4 +2,4 @@
 #include <iostream>
 #include "action_moving.h"
 
-Moving::Moving(std::array<int8_t, 2> position) : position(position) {}
+Moving::Moving(std::array<uint32_t, 2> position) : position(position) {}
